+++ include/Camera.h
@@ -11,7 +11,7 @@ class Camera {
 	public:
 		static void Follow(GameObject * newFocus);
 		static void Unfollow();
-		static void Update(float dt);
+		static void Update(double dt);
 
 		static Vec2 pos;
 		static Vec2 speed;
