+++ include/Vec2.h
@@ -5,17 +5,17 @@ using namespace std;
 
 class Vec2 {
 	public:
-		float x;
-		float y;
+		double x;
+		double y;
 
 		Vec2();
-		Vec2(float x, float y);
+		Vec2(double x, double y);
 
-		float GetX();
-		float GetY();
-		Vec2 Translate(Vec2 vector, float x, float y);
+		double GetX();
+		double GetY();
+		Vec2 Translate(Vec2 vector, double x, double y);
 		Vec2 Rotate(Vec2 vector, double angle);
-		Vec2 Rotate(Vec2 vector, double angle, float x, float y);
+		Vec2 Rotate(Vec2 vector, double angle, double x, double y);
 };
 
 #endif
